requirements.txt
# Core dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.12.0

# Optional: для расширенной функциональности
aiohttp>=3.9.0
